requirements.txt
BeautifulSoup==3.2.1
Django==1.6.4
PyYAML==3.11
South==0.8.4
Twisted==13.2.0
argparse==1.2.1
distribute==0.6.34
django-braces==1.4.0
django-grappelli==2.5.3
django-post-office==0.8.4
gevent==1.0.1
greenlet==0.4.2
jsonfield==0.9.20
nltk==2.0.4
psycopg2==2.5.2
python-dateutil==2.2
pytz==2014.2
requests==2.2.1
see==1.0.1
six==1.6.1
twitter==1.14.3
uWSGI==2.0.6
wsgiref==0.1.2
zope.interface==4.1.1
